+++ Source/Foundation/bsfUtility/Utility/BsSmallVector.h
@@ -12,7 +12,7 @@ namespace bs
 	 * Dynamically sized container that statically allocates enough room for @p N elements of type @p Type. If the element
 	 * count exceeds the statically allocated buffer size the vector falls back to general purpose dynamic allocator.
 	 */
-	template <class Type, u32 N>
+	template <class Type, u64 N>
 	class SmallVector final
 	{
 	public:
@@ -43,7 +43,7 @@ namespace bs
 				*this = std::move(other);
 		}
 
-		SmallVector(u32 size, const Type& value = Type())
+		SmallVector(u64 size, const Type& value = Type())
 		{
 			Append(size, value);
 		}
@@ -75,8 +75,8 @@ namespace bs
 			if(this == &other)
 				return *this;
 
-			u32 mySize = Size();
-			const u32 otherSize = other.Size();
+			u64 mySize = Size();
+			const u64 otherSize = other.Size();
 
 			// Use assignment copy if we have more elements than the other array, and destroy any excess elements
 			if(mySize > otherSize)
@@ -134,8 +134,8 @@ namespace bs
 			// vector
 			else
 			{
-				u32 mySize = Size();
-				const u32 otherSize = other.Size();
+				u64 mySize = Size();
+				const u64 otherSize = other.Size();
 
 				// Use assignment copy if we have more elements than the other array, and destroy any excess elements
 				if(mySize > otherSize)
@@ -172,8 +172,8 @@ namespace bs
 
 		SmallVector<ValueType, N>& operator=(std::initializer_list<Type> list)
 		{
-			u32 mySize = Size();
-			const u32 otherSize = (u32)list.size();
+			u64 mySize = Size();
+			const u64 otherSize = (u64)list.size();
 
 			// Use assignment copy if we have more elements than the list, and destroy any excess elements
 			if(mySize > otherSize)
@@ -240,14 +240,14 @@ namespace bs
 			return !(*this < other);
 		}
 
-		Type& operator[](u32 index)
+		Type& operator[](u64 index)
 		{
 			B3D_ASSERT(index < mSize && "Array index out-of-range.");
 
 			return mElements[index];
 		}
 
-		const Type& operator[](u32 index) const
+		const Type& operator[](u64 index) const
 		{
 			B3D_ASSERT(index < mSize && "Array index out-of-range.");
 
@@ -280,9 +280,9 @@ namespace bs
 
 		ConstReverseIterator Crend() const { return ConstReverseIterator(Begin()); }
 
-		u32 Size() const { return mSize; }
+		u64 Size() const { return mSize; }
 
-		u32 Capacity() const { return mCapacity; }
+		u64 Capacity() const { return mCapacity; }
 
 		Type* Data() { return mElements; }
 
@@ -333,7 +333,7 @@ namespace bs
 			typename = std::enable_if_t<std::is_convertible_v<typename std::iterator_traits<IteratorType>::iterator_category, std::input_iterator_tag>>>
 		void Append(IteratorType start, IteratorType end)
 		{
-			const u32 count = (u32)std::distance(start, end);
+			const u64 count = (u64)std::distance(start, end);
 
 			if((Size() + count) > Capacity())
 				this->Grow(Size() + count);
@@ -342,7 +342,7 @@ namespace bs
 			mSize += count;
 		}
 
-		void Append(u32 count, const Type& element)
+		void Append(u64 count, const Type& element)
 		{
 			if((Size() + count) > Capacity())
 				this->Grow(Size() + count);
@@ -397,14 +397,14 @@ namespace bs
 			return current;
 		}
 
-		void Remove(u32 index)
+		void Remove(u64 index)
 		{
 			Erase(Begin() + index);
 		}
 
 		bool Contains(const Type& element)
 		{
-			for(u32 i = 0; i < mSize; i++)
+			for(u64 i = 0; i < mSize; i++)
 			{
 				if(mElements[i] == element)
 					return true;
@@ -415,7 +415,7 @@ namespace bs
 
 		void RemoveValue(const Type& element)
 		{
-			for(u32 i = 0; i < mSize; i++)
+			for(u64 i = 0; i < mSize; i++)
 			{
 				if(mElements[i] == element)
 				{
@@ -427,31 +427,31 @@ namespace bs
 
 		void Clear()
 		{
-			for(u32 i = 0; i < mSize; ++i)
+			for(u64 i = 0; i < mSize; ++i)
 				mElements[i].~Type();
 
 			mSize = 0;
 		}
 
-		void Reserve(u32 capacity)
+		void Reserve(u64 capacity)
 		{
 			if(capacity > mCapacity)
 				Grow(capacity);
 		}
 
-		void Resize(u32 size, const Type& value = Type())
+		void Resize(u64 size, const Type& value = Type())
 		{
 			if(size > mCapacity)
 				Grow(size);
 
 			if(size > mSize)
 			{
-				for(u32 i = mSize; i < size; i++)
+				for(u64 i = mSize; i < size; i++)
 					new(&mElements[i]) Type(value);
 			}
 			else
 			{
-				for(u32 i = size; i < mSize; i++)
+				for(u64 i = size; i < mSize; i++)
 					mElements[i].~Type();
 			}
 
@@ -477,8 +477,8 @@ namespace bs
 		ConstReverseIterator crbegin() const { return Crbegin(); } // NOLINT
 		ConstReverseIterator crend() const { return Crend(); } // NOLINT
 
-		u32 size() const { return Size(); } // NOLINT
-		u32 capacity() const { return Capacity(); } // NOLINT
+		u64 size() const { return Size(); } // NOLINT
+		u64 capacity() const { return Capacity(); } // NOLINT
 
 		Type* data() { return Data(); } // NOLINT
 		const Type* data() const { return Data(); } // NOLINT
@@ -493,14 +493,14 @@ namespace bs
 		Iterator erase(ConstIterator start, ConstIterator end) { return Erase(start, end); } // NOLINT
 
 		void clear() { Clear(); } // NOLINT
-		void reserve(u32 capacity) { Reserve(capacity); } // NOLINT
-		void resize(u32 size, const Type& value = Type()) { Resize(size, value); } // NOLINT
+		void reserve(u64 capacity) { Reserve(capacity); } // NOLINT
+		void resize(u64 size, const Type& value = Type()) { Resize(size, value); } // NOLINT
 
 	private:
 		/** Returns true if the vector is still using its static memory and hasn't made any dynamic allocations. */
 		bool IsSmall() const { return mElements == (Type*)mStorage; }
 
-		void Grow(u32 capacity)
+		void Grow(u64 capacity)
 		{
 			B3D_ASSERT(capacity > N);
 
@@ -529,8 +529,8 @@ namespace bs
 		std::aligned_storage_t<sizeof(Type), alignof(Type)> mStorage[N];
 		Type* mElements = (Type*)mStorage;
 
-		u32 mSize = 0;
-		u32 mCapacity = N;
+		u64 mSize = 0;
+		u64 mCapacity = N;
 	};
 
 	/** @} */
